requirements.txt
Flask>=3.0,<4
Flask-SQLAlchemy>=3.1,<4
Flask-Login>=0.6,<1
Flask-WTF>=1.2,<2
email-validator>=2.1,<3
Werkzeug>=3.0,<4
Pillow>=10.0,<12
